Adding Bill Of Material.
All the individual microservices should point to this super POM (called BOM).
In here there are a set of dependencies and configurations common to all the services, such as spring boot, DB connectors, Jackson and so on.

This will allow us to upgrade one dependency in a centralised place and this will be cascaded down to all the individual services.